🔨 refactor(index.js): refactor VueFootprintsMixin and useFootprints functions The VueFootprintsMixin and useFootprints functions have been refactored to improve readability and maintainability. The e function has been extracted from the VueFootprintsMixin and is now used in the computed property of the useFootprints function. This change improves code reuse and reduces duplication.
🎉 feat(composables.js, core.js, index.js, mixin.js): add support for footprints
This commit adds support for footprints, which are a way to track the user's navigation history. The `useFootprints` composable is added to retrieve the footprints for a given route and options. The `getFootprints` function is added to the `core.js` file to calculate the footprints based on the matched routes and options. The `VueFootprintsMixin` is added to the `mixin.js` file to provide a global `$footprints` computed property that can be used in any component. Finally, the `index.js` file is updated to export the `VueFootprintsMixin` and `useFootprints` for use in other parts of the application.